+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Allow customization of login page's CSS. Refs STCOR-643.
 * Allow customization of navbar CSS. Refs STCOR-644.
 * Add `cs_CZ` (Czech, Czechia) to the supported locales. Refs STCOR-645.
+* Optionally display Okapi env, mod-configuration, or stripes.config values on about page. Refs STCOR-603.
 
 ## [8.2.0](https://github.com/folio-org/stripes-core/tree/v8.2.0) (2022-06-14)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v8.1.0...v8.2.0)